refactor: made changes in light of confirmic addition
